Work on rtunes algo continues; it seems to work with the 0 or 1 (unplot or plot) passed in as a uint16_t. Switching it to a uint8_t produces an odd crash when using -O2. It's working out slower than my original table approach so I'm not sure it's worth chasing up. Can't quite help myself though.

+++ rtunes.h
@@ -22,6 +22,6 @@
 
 #include <stdint.h>
 
-void rtunes_pixel( uint8_t, uint8_t, uint8_t ) __preserves_regs(iyl,iyh) __z88dk_callee;
+void rtunes_pixel( uint8_t, uint8_t, uint16_t ) __preserves_regs(iyl,iyh) __z88dk_callee;
 
 #endif
